requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.8.2
pydantic-settings==2.4.0
Pillow==10.3.0
PyYAML==6.0.2
